V1.2 - Refactored utilility modules from app_route_1_2.py to util.py - Added jquery div auto refresh on PrintingRequestFiles.html
