+++ PSITSweb/templates/app_templates_1_2/PrintingRequest.html
@@ -37,7 +37,7 @@
             </span>
             <label for="idnumber">Student Information  </label><br><input id="idnumber" class="file_input" type="number" name="idnum" placeholder="ID NUMBER" pattern="^[0-9]{8,}" required>
             
-            <br><br><label for="f2p">Files to print</label><br>
+            <br><br><label for="f2p">Files to print <p style="color: red;">(upload multiple files)</p></label><br>
             <input id="f2p" class="file_input" type="file" name="files[]" accept=".docx,.doc,.pdf,.xls" multiple="" required>
             <br><br>
             <button class="normalButton" type="submit">SUBMIT</button>
